requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
#click @ file:///opt/concourse/worker/volumes/live/2d66025a-4d79-47c4-43be-6220928b6c82/volume/click_1646056610594/work
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.2
Flask==2.2.2
#Flask @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8am4k38yjk/croots/recipe/flask_1658870771936/work
fonttools==4.37.3
gunicorn==20.1.0
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
#itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
jedi==0.18.1
#Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
#MarkupSafe @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d4a9444f-bd4c-4043-b47d-cede33979b0fve7bm42r/croots/recipe/markupsafe_1654597878200/work
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-pretrained-vit==0.0.7
pyzmq==24.0.0
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tinycss2==1.1.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cpu
torchvision==0.13.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
#Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
widgetsnbextension==4.0.3
#python==3.10.4